pyproject.toml
[project]
name = "policylab"
version = "1.0.0"
description = "SwarmCast — multi-agent behavioral simulation with LLM swarm intelligence for macro scenario analysis"
authors = [{ name = "Ambar" }]
requires-python = ">=3.10"
dependencies = [
    "gdm-concordia>=2.4.0",
    "openai>=1.0.0",
    "numpy",
    "sentence-transformers",
    "plotly>=5.0.0",
    "reportlab>=4.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
policylab = "policylab.cli:main"
